requirements.txt
streamlit
pandas
azure-storage-blob
msal